Move ScabbardMessage to protocol::v3
