+++ services/scabbard/libscabbard/src/protocol/v3/message.rs
